requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor-5==0.2.14
django-cors-headers==4.5.0
django-modeltranslation==0.19.10
django-restframework==0.0.1
django-tinymce==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.2.0
sqlparse==0.5.1
uritemplate==4.1.1
